fix: Use a different algorithm for generating a preview
That was not a good algorithm
The reason is this: suppose you have a set of 10 reps and 10 kg, a set of 10 reps and 20 kg, and another set of 10 reps and 10 kg.
This would be displayed as "2 × 10 reps 10 kg, 10 reps 20 kg". Sets shouldn't be grouped together if they aren't consecutive, so that example should be displayed as "10 reps 10 kg, 10 reps 20 kg, 10 reps 10 kg" (which it now is)